requirements.txt
requests==2.31.0
python-dotenv==0.19.1
agentops==0.2.0
chromadb==0.5.0
llama-index-core==0.10.42
llama-index-legacy==0.9.48